feat: 유저 상태 스토어 작성
